requirements.txt
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
scikit-image==0.18.0
scipy==1.5.4
torch==1.9.1
torchvision==0.10.1
